Create Caesar.java
This is my first code of my own solution to the Coding of Caesar task. This code is original because "if" statement is used only once. I made a formula for encoding and decoding. This formula seems complicated, but creating it was very interesting to me. I made two different formulas in these programs: Caesar.java and Caesar1.java.